move data loaders to their domain directories
